requirements.txt
pandas==1.5.3
requests==2.28.1
python-dotenv==1.0.0
